Fix remaining internals: Ast Index UInt, Bound AssocTypeArg.assoc.ty + LifetimeParam, Check segments.toList, De forward declarations
- Ast.kt: Index(dstFields.size.toUInt(), Span.callSite()) — Index takes UInt
- Bound.kt: arg.assoc.ty for AssocTypeArg, def directly as GenericParam (not GenericParam.Lifetime(def))
- Check.kt: !params.isEmpty() instead of isNotEmpty(), segments.toList().lastOrNull()
- De.kt: forward declarations for deserializeCustomIdentifier/StructInPlace/TupleInPlace
- Error count: 3526 (internals nearly clean: Ast 1, Bound 1, Check 1, De 7)

+++ serde_derive/src/commonMain/kotlin/io/github/kotlinmania/serderive/internals/Ast.kt
@@ -191,7 +191,7 @@ private fun fieldsFromAst(
         val member = if (ident != null) {
             Member.Named(ident.deepCopy())
         } else {
-            Member.Unnamed(io.github.kotlinmania.syn.Index(dstFields.size, io.github.kotlinmania.procmacro2.Span.callSite()))
+            Member.Unnamed(io.github.kotlinmania.syn.Index(dstFields.size.toUInt(), io.github.kotlinmania.procmacro2.Span.callSite()))
         }
         dstFields.add(
             Field(
